Move truncated diff to utils
