+++ src/components/HistoryTable.js
@@ -4,57 +4,28 @@ import type { RecordData, ResourceHistoryEntry, RouteLocation } from "../types";
 import type { Location } from "react-router-dom";
 
 import React, { PureComponent } from "react";
-import { diffJson } from "diff";
 
 import * as NotificationActions from "../actions/notifications";
-import { timeago, humanDate, parseHistoryFilters } from "../utils";
+import { diffJson, timeago, humanDate, parseHistoryFilters } from "../utils";
 import AdminLink from "./AdminLink";
 import Spinner from "./Spinner";
 import PaginatedTable from "./PaginatedTable";
 import { getClient } from "../client";
 import { omit, sortHistoryEntryPermissions } from "../utils";
 
-// Number of lines to show above/below changes in diffs.
-const DIFF_CONTEXT = 3;
-
 function Diff({ source, target }) {
-  const diff = diffJson(target, source);
+  const diff = diffJson(source, target);
   return (
     <pre className="json-record">
-      {diff.map((chunk, i) => {
-        const isFirstChunk = i == 0;
-        const isLastChunk = i == diff.length - 1;
-        const className = chunk.added
+      {diff.map((chunk: string, i) => {
+        const className = chunk.startsWith("+")
           ? "added"
-          : chunk.removed
+          : chunk.startsWith("-")
           ? "removed"
           : "";
-
-        let lines = chunk.value.split("\n").filter(part => part !== "");
-
-        // Truncate beginning of first chunk if larger than context.
-        if (isFirstChunk && lines.length > DIFF_CONTEXT) {
-          lines = ["..."].concat(lines.slice(-DIFF_CONTEXT));
-          // Truncate end of last chunk if larger than context.
-        } else if (isLastChunk && lines.length > DIFF_CONTEXT) {
-          lines = lines.slice(0, DIFF_CONTEXT).concat(["..."]);
-          // Truncate middle chunk only if larger than twice the context (above + below).
-        } else if (lines.length > DIFF_CONTEXT * 2) {
-          lines = lines
-            .slice(DIFF_CONTEXT)
-            .concat(["..."])
-            .concat(lines.slice(-DIFF_CONTEXT));
-        }
-
-        const prefixedChunk = lines
-          .map(line => {
-            const prefix = chunk.added ? "+ " : chunk.removed ? "- " : "  ";
-            return prefix + line;
-          })
-          .join("\n");
         return (
           <div key={i} className={className}>
-            <code>{prefixedChunk}</code>
+            <code>{chunk}</code>
           </div>
         );
       })}
